+++ evm/package.json
@@ -4,24 +4,40 @@
   "author": "Wormhole Contributors",
   "license": "Apache-2.0",
   "main": "./dist/cjs/index.js",
-  "types":"./dist/cjs/index.d.ts",
+  "types": "./dist/cjs/index.d.ts",
   "module": "./dist/esm/index.js",
+  "files": [
+    "dist/cjs",
+    "dist/esm"
+  ],
   "scripts": {
     "build:esm": "tsc -p tsconfig.esm.json",
     "build:cjs": "tsc -p tsconfig.cjs.json",
     "build": "npm run build:esm && npm run build:cjs",
     "generate": "typechain --target=ethers-v5 --out-dir=ts/src/types out/*/*.json",
-    "clean":"rm -rf dist && rm -rf node_modules"
+    "clean": "rm -rf dist && rm -rf node_modules"
+  },
+  "exports": {
+    ".": {
+      "import": "./dist/esm/index.js",
+      "require": "./dist/cjs/index.js",
+      "types": "./dist/cjs/index.d.ts"
+    },
+    "./*": {
+      "import": "./dist/esm/*/index.js",
+      "require": "./dist/cjs/*/index.js",
+      "types": "./dist/cjs/*/index.d.ts"
+    }
   },
   "dependencies": {
-    "@wormhole-foundation/sdk-base":"^0.7.0-beta.4",
-    "@wormhole-foundation/sdk-definitions":"^0.7.0-beta.4",
-    "@wormhole-foundation/sdk-evm":"^0.7.0-beta.4",
-    "@wormhole-foundation/example-liquidity-layer-definitions":"0.0.1",
+    "@wormhole-foundation/sdk-base": "^0.7.0-beta.4",
+    "@wormhole-foundation/sdk-definitions": "^0.7.0-beta.4",
+    "@wormhole-foundation/sdk-evm": "^0.7.0-beta.4",
+    "@wormhole-foundation/example-liquidity-layer-definitions": "0.0.1",
     "ethers": "^5.7.2"
   },
   "devDependencies": {
-    "envfile": "^6.18.0",
+    "envfile": "^7.1.0",
     "@typechain/ethers-v5": "^10.2.0",
     "@types/chai": "^4.3.4",
     "@types/mocha": "^10.0.1",
